[logstash-output] configurable+multiple field definitions in dry run
