+++ integration/logstash-plugins/logstash-output-vespa/build.gradle
@@ -63,6 +63,8 @@ dependencies {
     implementation 'com.fasterxml.jackson.core:jackson-core:2.18.2'
     implementation 'com.fasterxml.jackson.core:jackson-annotations:2.18.2'
 
+    implementation "org.yaml:snakeyaml:${snakeYamlVersion}"
+
     testImplementation 'org.jruby:jruby-complete:9.4.3.0'
     testImplementation 'org.jruby:jruby-core:9.4.3.0'
     testImplementation "org.apache.logging.log4j:log4j-core:${log4jVersion}"
