+++ integration/logstash-plugins/logstash-output-vespa/src/main/resources/application_package_template/schemas/document_type.sd
@@ -3,4 +3,5 @@ schema document_type {
     document document_type {
         # Fields will be added here
     }
+    # synthetic fields can be added here
 }
